Adding The ReportBack Android Project
A project developed by Chamara S, Rajtha D & Dilan D as part of a
Hackathon initiative conducted by Virtusa PVT LTD in 2012. The initial
idea for the project was from Chamara.
This application enables the user to record an audio clip while taking
pictures and afterwards allows the user to email that captured data as a
zip file to a predefined mail address. This is mainly aimed for
reporters on the move who don't have good mobile connectivity to send
their reports back to base.